+++ lib/rspec/core/deprecation.rb
@@ -21,7 +21,23 @@ module RSpec
     #
     # Used internally to print deprecation warnings
     def warn_deprecation(message)
-      warn message
+      RSpec.configuration.deprecation_io.puts(message)
     end
+
+    # @private
+    #
+    # Used internally to print the count of deprecation warnings
+    def deprecations?
+      RSpec.configuration.deprecation_io.deprecations > 0
+    end
+
+    # @private
+    #
+    # Used internally to print deprecation summary
+    def deprecations_summary
+      io = RSpec.configuration.deprecation_io
+      "There were #{io.deprecations} deprecations logged to #{io.description}"
+    end
+
   end
 end