+++ spec/rspec/core/configuration_spec.rb
@@ -16,6 +16,43 @@ module RSpec::Core
       end
     end
 
+    describe '#deprecation_io' do
+      it 'defaults to standard error' do
+        expect(config.deprecation_io.io).to eq $stderr
+      end
+
+      it 'is configurable' do
+        io = double 'deprecation io'
+        config.deprecation_io = io
+        expect(config.deprecation_io.io).to eq io
+      end
+
+      it 'is used by warn_deprecation' do
+        message = double
+        RSpec.configuration.deprecation_io.should_receive(:puts).with(message)
+        RSpec.warn_deprecation message
+      end
+    end
+
+    describe '#log_deprecations_to_file' do
+      let(:name) { 'filename.txt' }
+
+      around do |example|
+        config.log_deprecations_to_file name
+        example.run
+        File.delete name if File.exist? name
+      end
+
+      it 'sets the deprecation io to be a file with the supplied name' do
+        file = config.deprecation_io.io
+        expect(file.path).to eq name
+      end
+
+      it 'sets the deprecation io description to be the supplied file name' do
+        expect(config.deprecation_io.description).to eq name
+      end
+    end
+
     describe "#setup_load_path_and_require" do
       include_context "isolate load path mutation"
 
@@ -1209,7 +1246,7 @@ module RSpec::Core
 
       context "with :alias => " do
         it "is deprecated" do
-          RSpec::should_receive(:warn).with(/deprecated/)
+          RSpec::should_receive(:warn_deprecation).with(/deprecated/)
           config.add_setting :custom_option
           config.add_setting :another_custom_option, :alias => :custom_option
         end